+++ test/test_indirect_vector.cpp
@@ -1222,7 +1222,7 @@ namespace
       DataNDC data(compare_data.begin(), compare_data.end());
       data.clear();
 
-      CHECK_EQUAL(data.size(), size_t(0UL));
+      CHECK_EQUAL(data.size(), 0UL);
     }
 
     //*************************************************************************
