+++ test/test_map.cpp
@@ -206,7 +206,7 @@ namespace
     {
       Data data;
 
-      CHECK(data.size() == size_t(0UL));
+      CHECK(data.size() == 0UL);
       CHECK(data.empty());
       CHECK(data.capacity() == MAX_SIZE);
       CHECK(data.max_size() == MAX_SIZE);
@@ -1058,7 +1058,7 @@ namespace
       Data data(compare_data.begin(), compare_data.end());
       data.clear();
 
-      CHECK(data.size() == size_t(0UL));
+      CHECK(data.size() == 0UL);
     }
 
     //*************************************************************************
@@ -1066,9 +1066,9 @@ namespace
     {
       const Data data(initial_data.begin(), initial_data.end());
 
-      CHECK(data.count("3") == size_t(1UL));
+      CHECK(data.count("3") == 1UL);
 
-      CHECK(data.count("A") == size_t(0UL));
+      CHECK(data.count("A") == 0UL);
     }
 
     //*************************************************************************
@@ -1078,9 +1078,9 @@ namespace
 
       const EMap data(initial_data.begin(), initial_data.end());
 
-      CHECK(data.count(Key("3")) == size_t(1UL));
+      CHECK(data.count(Key("3")) == 1UL);
 
-      CHECK(data.count(Key("A")) == size_t(0UL));
+      CHECK(data.count(Key("A")) == 0UL);
     }
 
     //*************************************************************************
