+++ test/test_vector_pointer_external_buffer.cpp
@@ -105,7 +105,7 @@ namespace
     {
       Data data(buffer1, SIZE);
 
-      CHECK_EQUAL(data.size(), size_t(0UL));
+      CHECK_EQUAL(data.size(), 0UL);
       CHECK(data.empty());
       CHECK_EQUAL(data.capacity(), SIZE);
       CHECK_EQUAL(data.max_size(), SIZE);
@@ -116,7 +116,7 @@ namespace
     {
       CData data(buffer1, SIZE);
 
-      CHECK_EQUAL(data.size(), size_t(0UL));
+      CHECK_EQUAL(data.size(), 0UL);
       CHECK(data.empty());
       CHECK_EQUAL(data.capacity(), SIZE);
       CHECK_EQUAL(data.max_size(), SIZE);
@@ -1175,9 +1175,9 @@ namespace
     }
 
     //*************************************************************************
+#include "etl/private/diagnostic_array_bounds_push.h"
     TEST_FIXTURE(SetupFixture, test_insert_position_value_excess)
     {
-#include "etl/private/diagnostic_array_bounds_push.h"
       const size_t INITIAL_SIZE     = SIZE;
       int INITIAL_VALUE       = 1;
 
@@ -1194,8 +1194,8 @@ namespace
       offset = data.size();
 
       CHECK_THROW(data.insert(data.begin() + offset, &INITIAL_VALUE), etl::vector_full);
-#include "etl/private/diagnostic_pop.h"
     }
+#include "etl/private/diagnostic_pop.h"
 
     //*************************************************************************
     TEST_FIXTURE(SetupFixture, test_emplace_position_value)
@@ -1488,7 +1488,7 @@ namespace
       Data data(compare_data.begin(), compare_data.end(), buffer1, SIZE);
       data.clear();
 
-      CHECK_EQUAL(data.size(), size_t(0UL));
+      CHECK_EQUAL(data.size(), 0UL);
     }
 
     //*************************************************************************
@@ -1499,7 +1499,7 @@ namespace
       CData data(compare_data.begin(), compare_data.end(), buffer1, SIZE);
       data.clear();
 
-      CHECK_EQUAL(data.size(), size_t(0UL));
+      CHECK_EQUAL(data.size(), 0UL);
     }
 
     //*************************************************************************
